Create BMI_C.py
For those unaware, a human's body constitution is measured by dividing their weight by their height squared. The returned value is known as a "Body Mass Index" or BMI. This program is meant to act as a simple BMI calculator. The user inputs a weight and height, and is returned a BMI value and a general grade classification based on it.