Create T_T.py
This was one of my first "major" scripts that I created, and I finally decided to upload it. The "Time Till" script is intended to be a basic date-countdown program. The user inputs the month, day, and year of the date they desire, and the program will give an estimeate of the time remaining unitl it occurs. I plan on "cleaning" it up soon; e.g., actually adding commments, because even I have no memory of this place.